# reference-apps/shared/openapi.yaml
openapi: 3.1.0
info:
  title: DevStack Core - Reference API
  description: |
    Reference implementation demonstrating infrastructure integration patterns
    for production-grade applications.

    This API showcases:
    - HashiCorp Vault integration for secrets management
    - Multi-database connectivity (PostgreSQL, MySQL, MongoDB)
    - Redis caching and cluster management
    - RabbitMQ messaging patterns
    - Comprehensive health checks
    - Prometheus metrics integration

    **Note:** This is a reference implementation for learning and development.
    Production deployments should add authentication, rate limiting, and other
    security controls as documented in SECURITY_ASSESSMENT.md.

  version: 1.0.0
  contact:
    name: Development Team
    url: https://github.com/NormB/devstack-core
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Code-First Implementation (FastAPI with decorators)
  - url: http://localhost:8001
    description: API-First Implementation (Generated from OpenAPI spec)
  - url: https://localhost:8443
    description: Code-First Implementation (HTTPS/TLS with Vault PKI)
  - url: https://localhost:8444
    description: API-First Implementation (HTTPS/TLS with Vault PKI)
...
paths:
...
  /examples/vault/secret/{service_name}:
    get:
      tags:
        - Vault Examples
      summary: Get Secret Example
      description: |-
        Example: Fetch a secret from Vault

        This shows how to retrieve credentials for any service stored in Vault.
        Response is cached for 5 minutes to reduce Vault API calls.

        Raises:
            VaultUnavailableError: If Vault is unreachable or returns an error
            ResourceNotFoundError: If the secret doesn't exist
      operationId: get_secret_example_examples_vault_secret__service_name__get
      parameters:
        - name: service_name
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 50
            pattern: ^[a-zA-Z0-9_-]+$
            description: Service name (alphanumeric, hyphens, underscores only)
            title: Service Name
          description: Service name (alphanumeric, hyphens, underscores only)
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecretResponse'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /examples/vault/secret/{service_name}/{key}:
    get:
      tags:
        - Vault Examples
      summary: Get Secret Key Example
      description: |-
        Example: Fetch a specific key from a secret

        Useful when you only need one field (like just the password).
        Response is cached for 5 minutes to reduce Vault API calls.

        Raises:
            VaultUnavailableError: If Vault is unreachable or returns an error
            ResourceNotFoundError: If the secret or key doesn't exist
      operationId: get_secret_key_example_examples_vault_secret__service_name___key__get
      parameters:
        - name: service_name
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 50
            pattern: ^[a-zA-Z0-9_-]+$
            description: Service name (alphanumeric, hyphens, underscores only)
            title: Service Name
          description: Service name (alphanumeric, hyphens, underscores only)
        - name: key
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 100
            pattern: ^[a-zA-Z0-9_-]+$
            description: Secret key name (alphanumeric, hyphens, underscores only)
            title: Key
          description: Secret key name (alphanumeric, hyphens, underscores only)
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecretKeyResponse'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /examples/cache/{key}:
    get:
      tags:
        - Cache Examples
      summary: Get Cache Value
      description: 'Example: Get a value from cache'
      operationId: get_cache_value_examples_cache__key__get
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 200
            pattern: ^[a-zA-Z0-9_:.-]+$
            description: 'Cache key (alphanumeric and: - _ : . only)'
            title: Key
          description: 'Cache key (alphanumeric and: - _ : . only)'
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheGetResponse'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
        - Cache Examples
      summary: Set Cache Value
      description: 'Example: Set a value in cache with optional TTL'
      operationId: set_cache_value_examples_cache__key__post
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 200
            pattern: ^[a-zA-Z0-9_:.-]+$
            description: 'Cache key (alphanumeric and: - _ : . only)'
            title: Key
          description: 'Cache key (alphanumeric and: - _ : . only)'
        - name: value
          in: query
          required: true
          schema:
            type: string
            minLength: 0
            maxLength: 10000
            description: Value to cache (max 10KB)
            title: Value
          description: Value to cache (max 10KB)
        - name: ttl
          in: query
          required: false
          schema:
            type: integer
            maximum: 86400
            minimum: 1
            description: Time to live in seconds (1s to 24h)
            title: Ttl
          description: Time to live in seconds (1s to 24h)
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheSetResponse'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - Cache Examples
      summary: Delete Cache Value
      description: 'Example: Delete a value from cache'
      operationId: delete_cache_value_examples_cache__key__delete
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 200
            pattern: ^[a-zA-Z0-9_:.-]+$
            description: 'Cache key (alphanumeric and: - _ : . only)'
            title: Key
          description: 'Cache key (alphanumeric and: - _ : . only)'
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheDeleteResponse'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /examples/messaging/publish:
    post:
      tags:
        - Messaging Examples
      summary: Publish Message
      description: 'Example: Publish a message to a queue'
      operationId: publish_message_examples_messaging_publish_post
      parameters:
        - name: queue_name
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 100
            pattern: ^[a-zA-Z0-9_.-]+$
            description: 'Queue name (alphanumeric and: - _ . only)'
            title: Queue Name
          description: 'Queue name (alphanumeric and: - _ . only)'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: Message payload (JSON object, max 1MB)
              title: Message
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagePublishResponse'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /examples/messaging/queue/{queue_name}/info:
    get:
      tags:
        - Messaging Examples
      summary: Get Queue Info
      description: 'Example: Get information about a queue'
      operationId: get_queue_info_examples_messaging_queue__queue_name__info_get
      parameters:
        - name: queue_name
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 100
            pattern: ^[a-zA-Z0-9_.-]+$
            description: 'Queue name (alphanumeric and: - _ . only)'
            title: Queue Name
          description: 'Queue name (alphanumeric and: - _ . only)'
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueInfoResponse'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    CacheDeleteResponse:
      properties:
        key:
          type: string
          title: Key
          description: Cache key
        deleted:
          type: boolean
          title: Deleted
          description: Whether key was deleted
        action:
          type: string
          title: Action
          description: Action performed
          default: delete
      type: object
      required:
        - key
        - deleted
      title: CacheDeleteResponse
      description: Response for cache delete operations
    CacheGetResponse:
      properties:
        key:
          type: string
          title: Key
          description: Cache key
        value:
          anyOf:
            - type: string
            - type: "null"
          title: Value
          description: Cached value (null if not found)
        exists:
          type: boolean
          title: Exists
          description: Whether key exists in cache
        ttl:
          anyOf:
            - type: integer
            - type: string
            - type: "null"
          title: Ttl
          description: Time to live in seconds or 'no expiration'
      type: object
      required:
        - key
        - exists
      title: CacheGetResponse
      description: Response for cache get operations
    CacheSetResponse:
      properties:
        key:
          type: string
          title: Key
          description: Cache key
        value:
          type: string
          title: Value
          description: Value that was set
        ttl:
          anyOf:
            - type: integer
            - type: "null"
          title: Ttl
          description: Time to live in seconds (null if no expiration)
        action:
          type: string
          title: Action
          description: Action performed
          default: set
      type: object
      required:
        - key
        - value
      title: CacheSetResponse
      description: Response for cache set operations
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    MessagePublishResponse:
      properties:
        queue:
          type: string
          title: Queue
          description: Queue name
        message:
          additionalProperties: true
          type: object
          title: Message
          description: Message that was published
        action:
          type: string
          title: Action
          description: Action performed
          default: published
      type: object
      required:
        - queue
        - message
      title: MessagePublishResponse
      description: Response for message publishing
    QueueInfoResponse:
      properties:
        queue:
          type: string
          title: Queue
          description: Queue name
        exists:
          type: boolean
          title: Exists
          description: Whether queue exists
        message_count:
          anyOf:
            - type: integer
            - type: "null"
          title: Message Count
          description: Number of messages in queue
        consumer_count:
          anyOf:
            - type: integer
            - type: "null"
          title: Consumer Count
          description: Number of consumers
      type: object
      required:
        - queue
        - exists
      title: QueueInfoResponse
      description: Response for queue information
    SecretKeyResponse:
      properties:
        service:
          type: string
          title: Service
          description: Service name
        key:
          type: string
          title: Key
          description: Secret key name
        value:
          anyOf:
            - type: string
            - type: "null"
          title: Value
          description: Secret value (sensitive values masked)
        note:
          type: string
          title: Note
          description: Additional information
      type: object
      required:
        - service
        - key
        - note
      title: SecretKeyResponse
      description: Response for specific secret key retrieval
    SecretResponse:
      properties:
        service:
          type: string
          title: Service
          description: Service name
        data:
          additionalProperties: true
          type: object
          title: Data
          description: Secret data (passwords masked)
        note:
          type: string
          title: Note
          description: Additional information
      type: object
      required:
        - service
        - data
        - note
      title: SecretResponse
      description: Response for secret retrieval
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API Key authentication (not currently implemented in reference app).
        Production deployments should implement API key validation.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication (not currently implemented in reference app).
        Production deployments should implement JWT validation with proper claims.